refactor(web): extract MCP language model selector for unit tests
