+++ packages/web/src/ee/features/mcp/askCodebase.ts
@@ -18,9 +18,9 @@ import { captureEvent } from "@/lib/posthog";
 import { createAudit } from "@/ee/features/audit/audit";
 import { createMessageStream } from "@/ee/features/chat/agent";
 import { getPromptCacheStrategy } from "@/ee/features/chat/promptCaching";
+import { selectConfiguredLanguageModel } from "@/ee/features/mcp/selectConfiguredLanguageModel";
 
 const logger = createLogger('ask-codebase-api');
-type ConfiguredLanguageModel = Awaited<ReturnType<typeof getConfiguredLanguageModels>>[number];
 
 export type AskCodebaseParams = {
     query: string;
@@ -35,64 +35,6 @@ export type AskCodebaseResult = {
     chatId: string;
     chatUrl: string;
     languageModel: LanguageModelInfo;
-};
-
-const formatLanguageModelName = (model: Pick<LanguageModelInfo, 'provider' | 'model'>) =>
-    `${model.provider}/${model.model}`;
-
-const formatConfiguredLanguageModelLabel = (model: Pick<LanguageModelInfo, 'provider' | 'model' | 'displayName'>) =>
-    model.displayName ? `'${model.displayName}'` : formatLanguageModelName(model);
-
-export const selectConfiguredLanguageModel = (
-    configuredModels: ConfiguredLanguageModel[],
-    requestedLanguageModel: Pick<LanguageModelInfo, 'provider' | 'model' | 'displayName'>
-): {
-    languageModelConfig?: ConfiguredLanguageModel;
-    error?: ServiceError;
-} => {
-    const candidateModels = configuredModels.filter(
-        (model) => model.provider === requestedLanguageModel.provider && model.model === requestedLanguageModel.model
-    );
-    if (candidateModels.length === 0) {
-        return {
-            error: {
-                statusCode: StatusCodes.BAD_REQUEST,
-                errorCode: ErrorCode.INVALID_REQUEST_BODY,
-                message: `Language model '${formatLanguageModelName(requestedLanguageModel)}' is not configured.`,
-            } satisfies ServiceError,
-        };
-    }
-
-    if (requestedLanguageModel.displayName) {
-        const matchingModel = candidateModels.find(
-            (model) => model.displayName === requestedLanguageModel.displayName
-        );
-        if (!matchingModel) {
-            const availableDisplayNames = candidateModels.map(formatConfiguredLanguageModelLabel).join(', ');
-            return {
-                error: {
-                    statusCode: StatusCodes.BAD_REQUEST,
-                    errorCode: ErrorCode.INVALID_REQUEST_BODY,
-                    message: `Language model '${formatLanguageModelName(requestedLanguageModel)}' is configured, but not with displayName '${requestedLanguageModel.displayName}'. Available matches: ${availableDisplayNames}.`,
-                } satisfies ServiceError,
-            };
-        }
-
-        return { languageModelConfig: matchingModel };
-    }
-
-    if (candidateModels.length > 1) {
-        const availableDisplayNames = candidateModels.map(formatConfiguredLanguageModelLabel).join(', ');
-        return {
-            error: {
-                statusCode: StatusCodes.BAD_REQUEST,
-                errorCode: ErrorCode.INVALID_REQUEST_BODY,
-                message: `Language model '${formatLanguageModelName(requestedLanguageModel)}' matches multiple configured models. Pass displayName to disambiguate. Available matches: ${availableDisplayNames}.`,
-            } satisfies ServiceError,
-        };
-    }
-
-    return { languageModelConfig: candidateModels[0] };
 };
 
 const blockStreamUntilFinish = async <T extends UIMessage<unknown, UIDataTypes, UITools>>(
